Added support for LittleFS for storing web server files
This update modifies the code to use the LittleFS library instead of SPIFFS for storing the web server files. LittleFS is a more efficient file system for the ESP8266 and is recommended by the manufacturer. The update includes changes to the file system initialization, file handling, and file paths to ensure that the web server files are stored in the appropriate directory in the LittleFS file system. The update also includes changes to the HTML file to ensure that the web server is served with the appropriate content. These changes improve the overall performance and reliability of the web server on the ESP8266.