Added design to HTML with a CSS files.
This update includes changes to the HTML and code files to create a design for the web interface. The HTML now includes a container div that centers the content and limits its width to 800px. The CSS styles have been adjusted to make the content more legible and easier to interact with on mobile devices. The background color has been changed to a dark shade of grey and the font color is white. The font size for the h1 header has been increased and the input button has been made larger and more prominent. Overall, these changes should improve the user experience for those accessing the interface on mobile devices.