Added JSON API functionality to the server
This code change adds a new feature to the server that allows clients to retrieve information in JSON format using a new JSON API. The API is accessible through the "/state" endpoint, and clients can send a GET request to retrieve the current state of the server. The state is returned in JSON format.

The new JSON API was implemented using the AsyncCallbackJsonWebHandler class, which allows for asynchronous handling of JSON requests. The code is designed to be scalable and easily extendable, allowing for the addition of new endpoints and functionality in the future.